+++ agent/train_agent.py
@@ -119,7 +119,7 @@ class TrainAgent:
         observation_sequence_lengths = torch.tensor([observation_sequence_length_queue.get()
                                                     for observation_sequence_length_queue
                                                     in self.__runner_observation_sequence_length_queues])
-        assert observation_sequence_lengths.dtype == torch.int
+        assert observation_sequence_lengths.dtype == torch.long
         actor_actions = self.__actor.forward_model(
             observations=observations,
             previous_actions=self.__buffer.last_actions(history_size=self.__history_size - 1),
